# backend/openapi/k12-api-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "K12 平台接口文档",
    "description": "K12 多智能体教学平台可调用接口。包含 IAM 账号安全、课程、智能体、结构化作业题目、答题与批改，统一通过 Gateway 和 JWT 鉴权。",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "网关服务"
    },
    {
      "url": "http://localhost:8081",
      "description": "用户与权限服务"
    },
    {
      "url": "http://localhost:8082",
      "description": "课程资源服务"
    },
    {
      "url": "http://localhost:8083",
      "description": "AI 智能体服务"
    },
    {
      "url": "http://localhost:8084",
      "description": "作业与批改服务"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/agents": {
      "get": {
        "tags": [
          "04-AI智能体服务"
        ],
        "summary": "智能体列表",
        "operationId": "listAgents",
        "description": "需要 ROLE_ADMIN 或 agent:read 权限。",
        "responses": {
          "200": {
            "description": "智能体列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListApiResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "04-AI智能体服务"
        ],
        "summary": "创建智能体",
        "operationId": "createAgent",
        "description": "需要 ROLE_ADMIN 或 agent:create 权限。",
        "requestBody": {
          "$ref": "#/components/requestBodies/AgentRequest"
        },
        "responses": {
          "201": {
            "description": "创建成功的智能体",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/agents/{id}": {
      "get": {
        "tags": [
          "04-AI智能体服务"
        ],
        "summary": "智能体详情",
        "operationId": "getAgent",
        "description": "需要 ROLE_ADMIN 或 agent:read 权限。",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "智能体详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentApiResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "04-AI智能体服务"
        ],
        "summary": "更新智能体",
        "operationId": "updateAgent",
        "description": "需要 ROLE_ADMIN 或 agent:update 权限。",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/AgentRequest"
        },
        "responses": {
          "200": {
            "description": "更新后的智能体",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "04-AI智能体服务"
        ],
        "summary": "删除智能体",
        "operationId": "deleteAgent",
        "description": "需要 ROLE_ADMIN 或 agent:delete 权限。",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OkEmpty"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/agents/{agentCode}/runs": {
      "post": {
        "tags": [
          "04-AI智能体服务"
        ],
        "summary": "运行智能体",
        "operationId": "createAgentRun",
        "description": "需要 ROLE_ADMIN 或 agent:invoke 权限。Java 从 JWT 获取用户 ID，调用 Python Runtime，并保存运行记录和产物。",
        "parameters": [
          {
            "name": "agentCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9-]{1,63}$"
            },
            "example": "study-plan"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunRequest"
              },
              "example": {
                "inputText": "为我制定八年级数学一次函数复习计划",
                "sessionId": "session-demo-001",
                "executionMode": "SYNC",
                "context": {
                  "grade": "八年级",
                  "durationMinutes": 40,
                  "weakPoints": [
                    "函数图像"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "SYNC 运行完成并已保存",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunApiResponse"
                }
              }
            }
          },
          "202": {
            "description": "ASYNC 运行已进入 RabbitMQ 队列，状态为 PENDING",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/v1/agents/runs": {
      "get": {
        "tags": [
          "04-AI智能体服务"
        ],
        "summary": "运行记录列表",
        "operationId": "listAgentRuns",
        "description": "需要 ROLE_ADMIN 或 agent:read 权限。普通用户只返回自己的运行记录，管理员可查看全部。",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "运行记录分页",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunPageApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/agents/runs/{runId}": {
      "get": {
        "tags": [
          "04-AI智能体服务"
        ],
        "summary": "运行记录详情",
        "operationId": "getAgentRun",
        "description": "需要 ROLE_ADMIN 或 agent:read 权限。普通用户只能查看自己的运行详情。",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          }
        ],
        "responses": {
          "200": {
            "description": "运行详情和全部产物",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/agents/runs/{runId}/cancel": {
      "post": {
        "tags": [
          "04-AI智能体服务"
        ],
        "summary": "取消异步任务",
        "operationId": "cancelAgentRun",
        "description": "需要 ROLE_ADMIN 或 agent:invoke。普通用户仅可取消自己的任务，管理员可取消全部。仅 ASYNC 的 PENDING/RUNNING 可取消，重复取消幂等。当前为业务取消：Java 拒绝迟到结果，不会中断 Python 进程或删除已入队消息。",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          }
        ],
        "responses": {
          "200": {
            "description": "任务已取消",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/v1/agents/runs/{runId}/retry": {
      "post": {
        "tags": [
          "04-AI智能体服务"
        ],
        "summary": "人工重试失败的异步任务",
        "operationId": "retryAgentRun",
        "description": "需要 ROLE_ADMIN 或 agent:invoke，且原记录对调用者可见。仅 FAILED/TIMED_OUT 的 ASYNC 可重试，重新校验智能体启用状态，复制输入创建新 runId。使用当前 JWT 身份，不冒用原用户。每次调用创建新任务，客户端不可自动重试此 POST；原任务可能仍在 Worker 执行，涉及外部写操作时需人工确认。",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          }
        ],
        "responses": {
          "202": {
            "description": "新的异步任务已受理",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/v1/agents/runs/{runId}/artifacts": {
      "get": {
        "tags": [
          "04-AI智能体服务"
        ],
        "summary": "运行产物列表",
        "operationId": "listAgentRunArtifacts",
        "description": "需要 ROLE_ADMIN 或 agent:read 权限。返回图表、表格、图片和文件等前端可渲染产物。",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          }
        ],
        "responses": {
          "200": {
            "description": "运行产物",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentArtifactListApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "example": 1
        }
      }
    },
    "requestBodies": {
...
      "AgentRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AgentRequest"
            },
            "example": {
              "code": "code-tutor",
              "name": "Code Tutor Agent",
              "type": "CODE_TUTOR",
              "description": "Guide students through programming exercises"
            }
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "未登录、JWT 无效或 JWT 已过期",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            },
            "example": {
              "code": 401,
              "message": "未登录、登录已过期或令牌无效",
              "data": null,
              "timestamp": "2026-08-14T02:30:00Z"
            }
          }
        }
      },
      "Forbidden": {
        "description": "已登录，但没有访问该资源的权限",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            },
            "example": {
              "code": 403,
              "message": "没有访问该资源的权限",
              "data": null,
              "timestamp": "2026-08-14T02:30:00Z"
            }
          }
        }
      },
      "BadRequest": {
        "description": "请求参数或业务输入不合法",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "唯一数据重复或资源状态冲突",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            }
          }
        }
      },
      "OkEmpty": {
        "description": "操作成功",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            }
          }
        }
      },
      "BadGateway": {
        "description": "Python Runtime 返回的数据不符合内部契约",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Python Runtime 当前不可用，失败运行记录仍会保存",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ApiResponseBase": {
        "type": "object",
        "required": [
          "code",
          "message",
          "timestamp"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EmptyApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "nullable": true
              }
            }
          }
        ]
      },
...
      "AgentRequest": {
        "type": "object",
        "required": [
          "code",
          "name",
          "type"
        ],
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9-]{1,63}$",
            "example": "code-tutor"
          },
          "name": {
            "type": "string",
            "maxLength": 128,
            "example": "Code Tutor Agent"
          },
          "type": {
            "type": "string",
            "maxLength": 64,
            "example": "CODE_TUTOR"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "example": "Guide students through programming exercises"
          }
        }
      },
      "Agent": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AgentRequest"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64",
                "example": 1
              },
              "status": {
                "type": "string",
                "example": "ENABLED"
              },
              "updatedTime": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "AgentApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/Agent"
              }
            }
          }
        ]
      },
      "AgentListApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        ]
      },
      "AgentRunRequest": {
        "type": "object",
        "required": [
          "inputText"
        ],
        "properties": {
          "inputText": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20000
          },
          "sessionId": {
            "type": "string",
            "maxLength": 64,
            "nullable": true
          },
          "executionMode": {
            "type": "string",
            "enum": [
              "SYNC",
              "ASYNC"
            ],
            "default": "SYNC"
          },
          "context": {
            "type": "object",
            "maxProperties": 50,
            "additionalProperties": true,
            "default": {}
          }
        }
      },
      "AgentArtifact": {
        "type": "object",
        "properties": {
          "artifactId": {
            "type": "string",
            "maxLength": 64
          },
          "kind": {
            "type": "string",
            "enum": [
              "CHART",
              "IMAGE",
              "TABLE",
              "TEXT",
              "FILE"
            ]
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "mimeType": {
            "type": "string"
          },
          "storageUri": {
            "type": "string",
            "nullable": true
          },
          "payload": {
            "nullable": true
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "checksumSha256": {
            "type": "string",
            "nullable": true
          },
          "createdTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "AgentRun": {
        "type": "object",
        "properties": {
          "runId": {
            "type": "string"
          },
          "agentCode": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "executionMode": {
            "type": "string",
            "enum": [
              "SYNC",
              "ASYNC"
            ]
          },
          "inputText": {
            "type": "string"
          },
          "inputContext": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "SUCCEEDED",
              "FAILED",
              "TIMED_OUT",
              "CANCELLED"
            ]
          },
          "outputText": {
            "type": "string",
            "nullable": true
          },
          "outputMetadata": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "durationMs": {
            "type": "integer",
            "format": "int64"
          },
          "startedTime": {
            "type": "string",
            "format": "date-time"
          },
          "finishedTime": {
            "type": "string",
            "format": "date-time"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "artifacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentArtifact"
            }
          }
        }
      },
      "AgentRunSummary": {
        "type": "object",
        "properties": {
          "runId": {
            "type": "string"
          },
          "agentCode": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "executionMode": {
            "type": "string",
            "enum": [
              "SYNC",
              "ASYNC"
            ]
          },
          "status": {
            "type": "string"
          },
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "durationMs": {
            "type": "integer",
            "format": "int64"
          },
          "startedTime": {
            "type": "string",
            "format": "date-time"
          },
          "finishedTime": {
            "type": "string",
            "format": "date-time"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AgentRunPage": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          },
          "hasNext": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentRunSummary"
            }
          }
        }
      },
      "AgentRunApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/AgentRun"
              }
            }
          }
        ]
      },
      "AgentRunPageApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/AgentRunPage"
              }
            }
          }
        ]
      },
      "AgentArtifactListApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AgentArtifact"
                }
              }
            }
          }
        ]
      },
...
    }
  }
}